Move trainer to contrib
